requirements.txt
numpy
matplotlib
hydra-core
omegaconf
ollama
